move CFM to filo.core.query
